Plexams.Generators.Schedule

Documentation

generateSchedule :: Plan -> (Plan, [AddExamToSlot])

scheduleExamsWithSameName :: Plan -> (Plan, [AddExamToSlot])

Schedule exams with same name in the same slot.

This only works if one of them is already planned. Ignores exams with same name that are planned in different slots.